Rename AddContactStep to EditContactNameStep
